refactored code, removed .text: bug
