+++ .gitignore
@@ -1,0 +1,2 @@
+bin/
+src/bochsrc.txt
