--- src/bochsrc.txt
@@ -1,10 +1,0 @@
-    megs:            32
-    display_library: sdl2
-    romimage:        file=/usr/share/bochs/BIOS-bochs-latest
-    vgaromimage:     file=/usr/share/bochs/VGABIOS-lgpl-latest
-    ata0-master:     type=cdrom, path=bin/os.iso, status=inserted
-    boot:            cdrom
-    log:             log/bochslog.txt
-    clock:           sync=realtime, time0=local
-    cpu:             count=1, ips=1000000c
-    com1:            enabled=1, mode=file, dev=com1.out
