+++ src/kmain.cpp
@@ -1,24 +1,21 @@
-#include "frame_buffer.h"
-#include "serial_port.cpp"
-#include "frame_buffer.cpp"
-/** fb_write_cell:
-     *  Writes a character with the given foreground and background to position i
-     *  in the framebuffer.
-     *
-     *  @param i  The location in the framebuffer
-     *  @param c  The character
-     *  @param fg The foreground color
-     *  @param bg The background color
-     */
+#include "output/frame_buffer.h"
+#include "output/serial_port.h"
+#include "output/frame_buffer.h"
 
 extern "C" int kmain(){
-      serial_configure(SERIAL_COM1_BASE, Baud_115200);
 
-  
-      char buffer[] = "t'is wankers innit";
+    //frame buffer test
+    serial_configure(SERIAL_COM1_BASE, Baud_115200);
+   
+    // max 14 characters
+    char buffer[] = "12345678901234";
 	    fb_write(buffer,sizeof(buffer),LIGHT_GREEN,BLACK);
-       
-      
-      return 0xcafebabe;
+     
+    //serial test
+    
+    char serial_buffer[] = "serial running";
+    serial_write(SERIAL_COM1_BASE, serial_buffer, sizeof(serial_buffer));
+
+    return 0xcafebabe;
 }
