+++ src/output/frame_buffer.h
@@ -54,6 +54,7 @@ void fb_write_cell(unsigned int i, char c, unsigned char fg, unsigned char bg);
  *  @param buf  Buffer that has contents to be written to screen
  *  @param len  Length of buffer
  */
-int fb_write(char *buf, unsigned int len);
+
+int fb_write(char *buf, unsigned int len, unsigned char fg, unsigned char bg);
 
 #endif /* INCLUDE_IO_H */
